ImGui事件实现。

+++ hazel/src/Hazel/Events/MouseEvent.h
@@ -54,7 +54,7 @@ namespace Hazel
     class HAZEL_API MouseButtonEvent : public Event
     {
     public:
-        inline float GetMouseButton() const { return m_Button; }
+        inline int GetMouseButton() const { return m_Button; }
 
         EVENT_CLASS_CATEGORY(EventCategoryMouse | EventCategoryInput)
 
@@ -86,8 +86,7 @@ namespace Hazel
     {
     public:
         MouseButtonReleasedEvent(int button)
-            : MouseButtonEvent(button) {
-        }
+            : MouseButtonEvent(button) {}
 
         std::string ToString() const override
         {
